refactored code into multiple packages
