+++ deploy/deployment.yaml
@@ -16,7 +16,7 @@ spec:
       serviceAccountName: argocd-interlace-controller
       containers:
         - name: argocd-interlace-controller
-          image: gcr.io/kg-image-registry/argocd-interlace-controller:dev42
+          image: gcr.io/kg-image-registry/argocd-interlace-controller:dev47
           imagePullPolicy: Always
           command:
             - argocd-interlace
