+++ pkg/controller/controller.go
@@ -1,30 +1,18 @@
 package controller
 
 import (
-	"bytes"
 	"context"
-	"crypto/tls"
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
-	"net/http"
-	"os"
-	"path/filepath"
 	"runtime/debug"
-	"strings"
 	"time"
 
-	"github.com/IBM/integrity-enforcer/enforcer/pkg/mapnode"
 	appv1 "github.com/argoproj/argo-cd/v2/pkg/apis/application/v1alpha1"
 	appClientset "github.com/argoproj/argo-cd/v2/pkg/client/clientset/versioned"
+	"github.com/gajananan/argocd-interlace/pkg/manifest"
 	"github.com/gajananan/argocd-interlace/pkg/utils"
-	"github.com/sigstore/k8s-manifest-sigstore/pkg/k8smanifest"
-	k8smnfutil "github.com/sigstore/k8s-manifest-sigstore/pkg/util"
+
 	log "github.com/sirupsen/logrus"
-	"github.com/tidwall/gjson"
-	"gopkg.in/yaml.v2"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	apiruntime "k8s.io/apimachinery/pkg/runtime"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/apimachinery/pkg/util/wait"
@@ -92,54 +80,13 @@ func newController(applicationClientset appClientset.Interface, namespace string
 			key, err := cache.MetaNamespaceKeyFunc(obj)
 
 			app, ok := obj.(*appv1.Application)
+
 			if ok {
-				appName := app.ObjectMeta.Name
-				//appPath := app.Status.Sync.ComparedTo.Source.Path
-				//appSourceRepoUrl := app.Status.Sync.ComparedTo.Source.RepoURL
-
-				desiredManifest := retriveApplicationResources(appName)
-
-				items := gjson.Get(desiredManifest, "items")
-
-				finalManifest := ""
-
-				log.Debug("len(items.Array()) ", len(items.Array()))
-
-				for i, item := range items.Array() {
-
-					targetState := gjson.Get(item.String(), "targetState").String()
-
-					finalManifest = prepareFinalManifest(targetState, finalManifest, i, len(items.Array())-1)
-				}
-				/*
-					fmt.Println("---------Event Recieved---------")
-					loc, _ := time.LoadLocation("UTC")
-					buildStartedOn := time.Now().In(loc)
-
-					fmt.Println()
-					fmt.Println("------------------ Source Git Repo  --------------")
-
-					fmt.Println("url: ", app.Status.Sync.ComparedTo.Source.RepoURL)
-					fmt.Println("path: ", app.Status.Sync.ComparedTo.Source.Path)
-					fmt.Println("targetRevision: ", app.Status.Sync.ComparedTo.Source.TargetRevision)
-					fmt.Println("commit id: ", app.Status.Sync.Revision)
-
-					appSourceRepoUrl := app.Status.Sync.ComparedTo.Source.RepoURL
-					appSourceRevision := app.Status.Sync.ComparedTo.Source.TargetRevision
-					appSourceCommitSha := app.Status.Sync.Revision
-
-					signAndGenerateProv(appName, appPath, appSourceRepoUrl, appSourceRevision, appSourceCommitSha,
-						finalManifest, imageRef, buildStartedOn)
-					fmt.Println("--------------------------------------------------")
-					fmt.Println("--------- Completed Processing Event---------")
-				*/
+				manifest.CreateEventHandler(app, PRIVATE_KEY_PATH, PUB_KEY_PATH)
 			}
 
 			if err == nil {
-				//log.Debug("Event received of type create for key [%s] ", key)
 				ctrl.appRefreshQueue.Add(key)
-				//log.Debug("Event queue size: %v", ctrl.appRefreshQueue.Len())
-
 			}
 
 		},
@@ -148,124 +95,14 @@ func newController(applicationClientset appClientset.Interface, namespace string
 				return
 			}
 			key, err := cache.MetaNamespaceKeyFunc(old)
-
 			oldApp, oldOK := old.(*appv1.Application)
 			newApp, newOK := new.(*appv1.Application)
 			if oldOK && newOK {
-				appName := newApp.ObjectMeta.Name
-				appPath := newApp.Status.Sync.ComparedTo.Source.Path
-
-				//log.Debug("oldApp.Status ", oldApp.Status)
-				//log.Debug("-------------------------")
-				//log.Debug("newApp.Status ", newApp.Status)
-
-				log.Debug(fmt.Sprintf("oldApp.Status.Health.Status %s ", oldApp.Status.Health.Status))
-
-				if oldApp.Status.OperationState != nil {
-					log.Debug(fmt.Sprintf("oldApp.Status.OperationState.Phase %s ", oldApp.Status.OperationState.Phase))
-
-				} else {
-					log.Debug(fmt.Sprintf("oldApp.Status.OperationState %s ", oldApp.Status.OperationState))
-				}
-
-				log.Debug(fmt.Sprintf("oldApp.Status.Sync.Status %s ", oldApp.Status.Sync.Status))
-
-				log.Debug(fmt.Sprintf("newApp.Status.Health.Status %s ", newApp.Status.Health.Status))
-
-				if newApp.Status.OperationState != nil {
-					log.Debug(fmt.Sprintf("newApp.Status.OperationState.Phase %s ", newApp.Status.OperationState.Phase))
-				} else {
-					log.Debug(fmt.Sprintf("newApp.Status.OperationState %s ", newApp.Status.OperationState))
-				}
-
-				log.Debug(fmt.Sprintf("newApp.Status.Sync.Status %s ", newApp.Status.Sync.Status))
-
-				if oldApp.Status.OperationState != nil &&
-					oldApp.Status.OperationState.Phase == "Running" &&
-					oldApp.Status.Sync.Status == "Synced" && //"Synced" &&
-					newApp.Status.OperationState != nil &&
-					newApp.Status.OperationState.Phase == "Running" &&
-					newApp.Status.Sync.Status == "OutOfSync" {
-
-					desiredManifest := retriveApplicationResources(appName)
-
-					items := gjson.Get(desiredManifest, "items")
-
-					finalManifest := ""
-					log.Debug("len(items.Array()) ", len(items.Array()))
-
-					diffCount := 0
-
-					appSourceRepoUrl := newApp.Status.Sync.ComparedTo.Source.RepoURL
-					appSourceRevision := newApp.Status.Sync.ComparedTo.Source.TargetRevision
-					appSourceCommitSha := newApp.Status.Sync.Revision
-					imageRef := getImageRef(appName, appPath, appSourceRepoUrl)
-
-					bundleYAMLBytes, err := getBundleManifest(imageRef)
-
-					manifestYAMLs := k8smnfutil.SplitConcatYAMLs(bundleYAMLBytes)
-
-					// if manifest not found, create it without check if diff exist
-					if err != nil {
-						diffCount += 1
-					}
-
-					for i, item := range items.Array() {
-
-						targetState := gjson.Get(item.String(), "targetState").String()
-
-						if diffCount == 0 {
-							diffExist := checkDiffWithBundle([]byte(targetState), manifestYAMLs)
-							if diffExist {
-								diffCount += 1
-							}
-						}
-
-						finalManifest = prepareFinalManifest(targetState, finalManifest, i, len(items.Array())-1)
-					}
-
-					log.Info("diffCount ", diffCount)
-					if diffCount > 0 {
-						log.Info("---------Event Recieved---------")
-						loc, _ := time.LoadLocation("UTC")
-						buildStartedOn := time.Now().In(loc)
-
-						log.Info()
-						log.Info("------------------ Source Git Repo  --------------")
-						/*
-							fmt.Println("url: ", oldApp.Status.Sync.ComparedTo.Source.RepoURL)
-							fmt.Println("path: ", oldApp.Status.Sync.ComparedTo.Source.Path)
-							fmt.Println("targetRevision: ", oldApp.Status.Sync.ComparedTo.Source.TargetRevision)
-							fmt.Println("commit id: ", oldApp.Status.Sync.Revision)
-							fmt.Println(oldApp.Status.History)
-							fmt.Println("----------")
-						*/
-
-						log.Info("url: ", newApp.Status.Sync.ComparedTo.Source.RepoURL)
-						log.Info("path: ", newApp.Status.Sync.ComparedTo.Source.Path)
-						log.Info("targetRevision: ", newApp.Status.Sync.ComparedTo.Source.TargetRevision)
-						log.Info("commit id: ", newApp.Status.Sync.Revision)
-
-						if finalManifest == "" {
-							log.Info("finalManifest is empty", finalManifest)
-						}
-						signAndGenerateProv(appName, appPath, appSourceRepoUrl, appSourceRevision, appSourceCommitSha,
-							finalManifest, imageRef, buildStartedOn)
-
-						log.Info("--------------------------------------------------")
-						log.Info("--------- Completed Processing Event---------")
-					}
-
-				}
-
+				manifest.UpdateEventHandler(oldApp, newApp, PRIVATE_KEY_PATH, PUB_KEY_PATH)
 			}
+
 			if err == nil {
-				log.Debug("Event received of type update for  [%s] ", key)
-
 				ctrl.appRefreshQueue.Add(key)
-
-				log.Debug("Event queue size: %v ", ctrl.appRefreshQueue.Len())
-
 			}
 
 		},
@@ -277,11 +114,7 @@ func newController(applicationClientset appClientset.Interface, namespace string
 
 			if err == nil {
 				log.Debug("Event received of type delete for key [%s] ", key)
-
 				//ctrl.appRefreshQueue.Add(key)
-
-				log.Debug("Event queue size %v", ctrl.appRefreshQueue.Len())
-
 			}
 
 		},
@@ -289,166 +122,6 @@ func newController(applicationClientset appClientset.Interface, namespace string
 
 	ctrl.informer = appInformer
 	return ctrl
-}
-
-func getImageRef(appName, appPath, appSourceRepoUrl string) string {
-	imageRef := ""
-	tokens := strings.Split(appSourceRepoUrl, "/")
-	if len(tokens) > 2 {
-		repoName := tokens[3]
-		imageRegistry := os.Getenv("IMAGE_REGISTRY")
-		imageName := fmt.Sprintf("%s-%s", repoName, appName)
-		tag := strings.ReplaceAll(appPath, "/", "-")
-		imageRef = fmt.Sprintf("%s/%s:%s", imageRegistry, imageName, tag)
-	}
-	return imageRef
-}
-func prepareFinalManifest(targetState, finalManifest string, counter int, numberOfitem int) string {
-	var obj *unstructured.Unstructured
-	err := json.Unmarshal([]byte(targetState), &obj)
-	if err != nil {
-	}
-	objBytes, _ := yaml.Marshal(obj)
-	endLine := ""
-	if !strings.HasSuffix(string(objBytes), "\n") {
-		endLine = "\n"
-	}
-	finalManifest = fmt.Sprintf("%s%s%s", finalManifest, string(objBytes), endLine)
-	finalManifest = strings.ReplaceAll(finalManifest, "object:\n", "")
-	if counter < numberOfitem {
-		finalManifest = fmt.Sprintf("%s---\n", finalManifest)
-	}
-	return finalManifest
-}
-
-func retriveApplicationResources(appName string) string {
-
-	baseUrl := os.Getenv("ARGOCD_API_BASE_URL")
-
-	desiredRscUrl := fmt.Sprintf("%s/%s/managed-resources", baseUrl, appName)
-
-	desiredManifest := queryAPI(desiredRscUrl, nil)
-
-	return desiredManifest
-}
-func signAndGenerateProv(appName, appPath, appSourceRepoUrl, appSourceRevision, appSourceCommitSha,
-	finalManifest, imageRef string, buildStartedOn time.Time) {
-
-	fmt.Println()
-	fmt.Println()
-	fmt.Println("------------------ Desired State Manifest --------------")
-	fmt.Println()
-
-	dirPath := filepath.Join("/tmp/output", appName, appPath)
-	if _, err := os.Stat(dirPath); os.IsNotExist(err) {
-		os.MkdirAll(dirPath, os.ModePerm)
-	}
-	fmt.Println(finalManifest)
-	outfilepath := filepath.Join(dirPath, "manifest.yaml")
-
-	utils.WriteToFile(string(finalManifest), outfilepath)
-
-	signManifest(outfilepath, imageRef, PRIVATE_KEY_PATH)
-
-	loc, _ := time.LoadLocation("UTC")
-	buildFinishedOn := time.Now().In(loc)
-
-	GenerateProvanance(appName, appPath, appSourceRepoUrl, appSourceRevision, appSourceCommitSha, PRIVATE_KEY_PATH, PUB_KEY_PATH, imageRef, buildStartedOn, buildFinishedOn)
-
-}
-
-func checkDiffWithBundle(targetObjYAMLBytes []byte, manifestYAMLs [][]byte) bool {
-
-	objNode, err := mapnode.NewFromBytes(targetObjYAMLBytes) // json
-	log.Debug("targetObjYAMLBytes ", string(targetObjYAMLBytes))
-	if err != nil {
-		log.Fatalf("objNode error from NewFromYamlBytes %s", err.Error())
-		// do somthing
-	}
-	found := false
-	for _, manifest := range manifestYAMLs {
-		//fmt.Println("manifest ", string(manifest))
-		mnfNode, err := mapnode.NewFromYamlBytes(manifest)
-		if err != nil {
-			log.Fatalf("mnfNode error from NewFromYamlBytes %s", err.Error())
-			// do somthing
-		}
-		diffs := objNode.Diff(mnfNode)
-
-		if diffs == nil || diffs.Size() == 0 {
-			found = true
-			break
-		}
-	}
-	return found
-
-}
-
-func getBundleManifest(imageRef string) ([]byte, error) {
-
-	image, err := k8smnfutil.PullImage(imageRef)
-
-	if err != nil {
-		log.Info("Error in pulling image err %s", err.Error())
-		return nil, err
-	}
-
-	imageManifest, _ := image.RawManifest()
-	log.Debug("imageManifest ", string(imageManifest))
-
-	concatYAMLbytes, err := k8smnfutil.GenerateConcatYAMLsFromImage(image)
-	if err != nil {
-		log.Info("Error in GenerateConcatYAMLsFromImage err %s", err.Error())
-		return nil, err
-	}
-	return concatYAMLbytes, nil
-}
-
-func signManifest(inputDir, imageRef, keyPath string) error {
-	so := &k8smanifest.SignOption{
-		ImageRef:         imageRef,
-		KeyPath:          keyPath,
-		Output:           "/tmp/output/manifest.signed",
-		UpdateAnnotation: true,
-		ImageAnnotations: nil,
-	}
-
-	_, err := k8smanifest.Sign(inputDir, so)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func queryAPI(url string, data map[string]string) string {
-
-	http.DefaultTransport.(*http.Transport).TLSClientConfig = &tls.Config{InsecureSkipVerify: true}
-	token := os.Getenv("ARGOCD_TOKEN")
-	var bearer = fmt.Sprintf("Bearer %s", token)
-	var dataJson []byte
-	if data != nil {
-		dataJson, _ = json.Marshal(data)
-	} else {
-		dataJson = nil
-	}
-	req, err := http.NewRequest("GET", url, bytes.NewBuffer(dataJson))
-	if err != nil {
-		log.Info("Error %s ", err)
-	}
-
-	req.Header.Add("Authorization", bearer)
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		log.Info("Error %s ", err)
-	}
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		log.Info("Error %s ", err)
-	}
-
-	return string([]byte(body))
 }
 
 func (c *controller) canProcessApp(obj interface{}) bool {
@@ -530,8 +203,6 @@ func (c *controller) processItem(key string) error {
 		return nil
 	}
 	//Use a switch clause instead and process the events based on the type
-
-	//log.Debug("argocd- has processed 1 event for object [%s]", obj)
 
 	return nil
 }
