+++ pkg/provenance/generate.go
@@ -1,4 +1,4 @@
-package controller
+package provenance
 
 import (
 	"crypto"
